Started fixing #876 and polishing (not done yet...)
