+++ src/server/graph-generation/biopax-metadata/generic-physical-entity-map.json
