+++ src/server/graph-generation/index.js
@@ -1,8 +1,7 @@
 const _ = require('lodash');
 const sbgn2CyJson = require('sbgnml-to-cytoscape');
 const pcServices = require('../pathway-commons');
-const {getBioPaxMetadata} = require('./biopax-metadata');
-const {getNodesGeneSynonyms} = require('./generic-physical-entities');
+const {getBioPaxMetadata, getNodesGeneSynonyms} = require('./biopax-metadata');
 
 //Get pathway name, description, and datasource
 //Requires a valid pathway uri
@@ -18,11 +17,15 @@ function getPathwayMetadata(uri) {
     get('Entity/comment').then(value => comments = value),
     get('Pathway/organism/displayName').then(value => organism = value)
   ]).then(() => ({ comments, dataSource, title, organism }));
+
 }
 
 /**
- * 
- * @param {*} uri URI representing the network
+ * Executes a Pathway Commons web query and builds a new JSON network model
+ * using the result SBGN (auto-converted to CyJSON) and additional metadata and data
+ * from the corresponding BioPAX L3 JSON-LD model and generic entities mapping file (json).
+ *
+ * @param {*} uri URI representing the network (query)
  * @returns A Cytoscape JSON which represents the network, enhanced with BioPAX metadata
  */
 function getPathwayElementJson(uri) {
@@ -54,7 +57,7 @@ function getPathwayElementJson(uri) {
 }
 
 /**
- * 
+ *
  * @param {*} uri URI representing the network
  * @returns A Cytoscape JSON which represents the network, enhanced with BioPAX metadata
  */
